--- src/main/java/cart/controller/ProductResponse.java
@@ -1,31 +1,0 @@
-package cart.controller;
-
-public class ProductResponse {
-    private final Long id;
-    private final String name;
-    private final Integer price;
-    private final String imageUrl;
-
-    public ProductResponse(Long id, String name, Integer price, String imageUrl) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Integer getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-}
